refactor: 컴포넌트 기준으로 폴더 구조 변경
