+++ src/components/Header/Header.jsx
@@ -1,4 +1,4 @@
-import "./css/Header.css";
+import "./Header.css";
 
 function Header() {
   return (
