+++ src/components/RestaurantDetailModal/RestaurantDetailModal.jsx
@@ -1,4 +1,4 @@
-import "./css/RestaurantModal.css";
+import "./RestaurantDetailModal.css";
 
 function RestaurantDetailModal() {
   return (
